\name{GraylingThesis-package}
\alias{GraylingThesis-package}
\alias{GraylingThesis}
\docType{package}
\title{Adaptive Design Methodology for Crossover Clinical Trials}
\description{The various functions used to carry out the work presented in the thesis
  are available. See the Appendix for further details. Error checking has been
  implemented and extensive testing performed to check for bugs. However, the
  functions should not be considered unbreakable.}
\details{
\tabular{ll}{
Package: \tab GraylingThesis\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-09-08\cr
License: \tab GPL-3\cr
}}
\author{Michael J. Grayling

Maintainer: Michael J. Grayling <mjg211@cam.ac.uk>
}
